Update patch set 1
Patch Set 1:

The singOut was a typo mistake. I will update it with signOut in next commit.

The another thing that I used sleep 40s as before using that function I am searching the model.  As after searching a model the page loading takes some time, If I used another function then at the same Xpath there is another element present in the DOM. If I didn’t use sleep then it doesn’t perform the action whatever we are expecting. As of now I didn’t find another way to handle this. If anything else I found I will update the script accordingly.

Regarding testing on particular browser we can test on both browser but not at the same time.
The ${browser} parameter we need to change as per the browser on which  want to execute the scripts.

Patch-set: 1
Reviewer: Gerrit User 108 <108@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0